Move websocket parsers to separate module
